% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayinfo.R
\name{arrayinfo}
\alias{arrayinfo}
\title{Obtain job ID and array information}
\usage{
arrayinfo(
  ID_env = "SLURM_ARRAY_TASK_ID",
  N_env = "SLURM_ARRAY_TASK_COUNT",
  MAX_env = "SLURM_ARRAY_TASK_MAX",
  MIN_env = "SLURM_ARRAY_TASK_MIN",
  jobID_env = "SLURM_JOB_ID",
  jobArrayID_env = "SLURM_ARRAY_JOB_ID",
  toint = TRUE
)
}
\arguments{
\item{ID_env}{the environment variable of the job array index value.}

\item{N_env}{the environment variable of the number of tasks in the job array}

\item{MAX_env}{the environment variable of the highest job array index value.}

\item{MIN_env}{the environment variable of the lowest job array index value}

\item{jobID_env}{the environment variable of the current job ID.}

\item{jobArrayID_env}{the environment variable of the first job ID of the array.}

\item{toint}{whether output the information as integer.}
}
\value{
the corresponding information.
}
\description{
Obtain job ID and array information in HPC (with SLURM, more see
[here](https://slurm.schedmd.com/job_array.html)).
}
\examples{
Sys.setenv("SLURM_ARRAY_TASK_ID"="123")
arrayinfo()$ID
}
