% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobinfo.R
\name{jobinfo}
\alias{jobinfo}
\title{Use HPC information to generate the seed and output file name}
\usage{
jobinfo(
  outfn_pre,
  seed = 1234,
  seedtype = "seed+ID",
  ID_env = "SLURM_ARRAY_TASK_ID",
  MAX_env = "SLURM_ARRAY_TASK_MAX",
  N_env = "SLURM_ARRAY_TASK_COUNT"
)
}
\arguments{
\item{outfn_pre}{the first part of the output file name.
The second part is the seed to be used.}

\item{seed}{the "default" seed to be used. Default to 1234. The specific seed
depends on \strong{seedtype}.}

\item{seedtype}{how the seed should be set. Default to  \strong{"seed+ID"}, i.e., the
sum of \code{seed} and the task ID from HPC. \strong{"ID"}, use the HPC task ID as the seed
directly. \strong{"seed"}, use the \code{seed} directly as the seed (note: if multiple
tasks are submitted simultaneously, the output file name may be the same).}

\item{ID_env}{the name of the system variable for task ID. Default to
"SLURM_ARRAY_TASK_ID".}

\item{MAX_env}{the name of the system variable for maximum task ID. Default
to "SLURM_ARRAY_TASK_MAX".}

\item{N_env}{the name of the system variable for the number of tasks. Default
to "SLURM_ARRAY_TASK_COUNT".}
}
\value{
a named list of information (length of 2). "seed" is a numeric value,
which can be used as the seed. "outfn" is the output file name.
}
\description{
Obtain the job array information from HPC
(e.g."SLURM_ARRAY_TASK_ID") to generate the seed and the output file name.
}
\examples{
Sys.setenv(SLURM_ARRAY_TASK_ID="5",
           SLURM_ARRAY_TASK_MAX="10",
           SLURM_ARRAY_TASK_COUNT="8")

out <- jobinfo("fit", seed=2022)
}
