% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_fit.R
\name{combine_fit}
\alias{combine_fit}
\title{Combine multiple files of fit objects}
\usage{
combine_fit(path = ".", pattern, outfn = NULL, ...)
}
\arguments{
\item{path}{path to the fit objects; will be use as path in list.files().}

\item{pattern}{wildcard used to identify the single fit objects; will be use
as pattern in list.files().}

\item{outfn}{the output file name of the combined fit objects. If outfn is
NULL (default), no file will be saved locally.}

\item{...}{other arguments to be used in list.files().}
}
\value{
a combined fit object.
}
\description{
combine the fit objects saved in separate files.
}
\examples{
\dontrun{
combine_fit("brmfit*", outfn="combined_brmfit")
}
}
